fix(facturacion): imp_neto/imp_iva/imp_total no truncan a peso entero (#1209)
El motor de facturación calcula el neto/IVA/total EXACTOS al centavo en
Decimal y esos son los valores que se le mandan a ARCA (armar_fecae) y se
codifican en el QR fiscal (armar_qr) para obtener el CAE. Pero al persistir
la fila en `facturas` se truncaban con int(round(float(...))), siguiendo por
error la convención de "enteros ARS" de la plata interna (contabilidad,
2026-06-07) — que no aplica a un documento fiscal cuyo importe legal ya
quedó fijado al centavo en ARCA. Una factura cuyo neto no fuera múltiplo de
100 (ej. $1001, IVA 21%=$210,21) quedaba impresa $0,21 por debajo de lo que
el CAE/QR realmente autorizaron.

Fix mínimo: imp_neto/imp_iva/imp_total pasan de INTEGER a NUMERIC(12,2)
(schema en dos capas: init_db() con guard idempotente + migración Alembic
h3i4j5k6l7m8) y engine.py deja de truncar en los 2 call-sites (emitir_factura
y emitir_nota_credito). pdf.py no necesitó cambios: _money/_plain ya
formatean con .2f. Como beneficio lateral, construir_comprobante_nc (que
reconstruye el importe de una NC desde el snapshot persistido de la factura
original) también queda corregido sin tocarlo.

Test nuevo (tests/test_facturacion_centavos.py) verifica que lo persistido
coincide exacto con lo mandado a ARCA y que el PDF muestra los centavos.

+++ backend/services/facturacion/repo.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 import json
 from dataclasses import dataclass
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Any, Optional
 
 
@@ -27,9 +28,10 @@ class Factura:
     doc_nro: str
     condicion_iva_receptor: int
     concepto: int
-    imp_neto: int
-    imp_iva: int
-    imp_total: int
+    # NUMERIC(12,2): Decimal exacto al centavo (bug #1209 — ver engine.py).
+    imp_neto: Decimal
+    imp_iva: Decimal
+    imp_total: Decimal
     moneda: str
     cliente_cuit: Optional[str]
     razon_social: Optional[str]
@@ -193,9 +195,9 @@ def insert_factura(
     doc_nro: str,
     condicion_iva_receptor: int,
     concepto: int,
-    imp_neto: int,
-    imp_iva: int,
-    imp_total: int,
+    imp_neto: Decimal,
+    imp_iva: Decimal,
+    imp_total: Decimal,
     moneda: str = "PES",
     cliente_cuit: Optional[str] = None,
     razon_social: Optional[str] = None,
